sparsha-backend/sparsha-backend/sparsha-web/Backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
google-generativeai>=0.3.0
googlemaps>=4.10.0
python-dotenv>=1.0.0